# specs/002-phase2-web/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: To-Do List API
  description: A simple API to manage a list of tasks.
  version: 1.0.0

paths:
  /tasks:
    get:
      summary: Get all tasks
      responses:
        '200':
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: The created task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '422':
          description: Validation Error

  /tasks/{task_id}:
    get:
      summary: Get a single task by its ID
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found.
    put:
      summary: Update a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: The updated task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found.
        '422':
          description: Validation Error
    delete:
      summary: Delete a task
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully.
        '404':
          description: Task not found.

components:
  schemas:
    TaskBase:
      type: object
      properties:
        title:
          type: string
          description: The main title of the task.
        description:
          type: string
          nullable: true
          description: A longer description for the task.
    TaskCreate:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          required:
            - title
    TaskUpdate:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          properties:
            completed:
              type: boolean
              description: Whether the task is completed.
    Task:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          properties:
            id:
              type: integer
              description: The unique identifier for the task.
            completed:
              type: boolean
              default: false
              description: Whether the task is completed.
          required:
            - id
            - title
            - completed